+++ c4j-acceptancetest/src/main/java/de/andrena/c4j/acceptancetest/lesson02/ASpecContract.java
@@ -14,7 +14,7 @@ public class ASpecContract implements ASpec {
 		if (post()) {
 			System.out.println("Post-Condition query");
 		}
-		return ignored();
+		return (Integer) ignored();
 	}
 
 	@Override
